AI-automated-Aspen-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
blockbuster==1.5.24
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
fireworks-ai==0.15.12
forbiddenfruit==0.1.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.21
langchain-core==0.3.51
langchain-fireworks==0.2.9
langchain-mcp-adapters==0.0.7
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.27
langgraph-api==0.0.48
langgraph-checkpoint==2.0.24
langgraph-cli==0.1.89
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.28
marshmallow==3.26.1
mcp==1.6.0
multidict==6.4.2
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.72.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
-e git+https://github.com/langchain-ai/react-agent.git@630b0ea3c3c98e4414343ee9044e25a650acc37c#egg=react_agent
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.1
structlog==25.2.0
tavily-python==0.5.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
